backend/requirements.txt
annotated-doc==0.0.4
bcrypt==4.0.1
passlib==1.7.4
PyJWT==2.10.1
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
fastapi==0.135.1
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
requests==2.32.5
soupsieve==2.8.3
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.41.0
